requirements.txt
# ============================================================
# Core runtime (API + common utilities)
# ============================================================
flask==3.0.3
python-dotenv
requests
httpx==0.27.2
numpy
pillow==10.2.0
tqdm==4.66.2
rapidfuzz==3.6.2
lxml==5.2.2
beautifulsoup4
playwright
selenium==4.18.1
scikit-image
scikit-learn
joblib
pandas==2.2.1
matplotlib
pillow-avif-plugin==1.4.3

# ============================================================
# Model/provider SDKs
# ============================================================
openai==1.13.3
anthropic
google-generativeai
boto3>=1.34.0

# ============================================================
# ML stack (used by multiple methods/evaluation)
# Note: for CUDA builds, install torch/torchvision per official instructions.
# ============================================================
torch
torchvision
transformers>=4.35
tokenizers>=0.14
huggingface_hub
accelerate
datasets

# ============================================================
# Method-specific dependencies
# ============================================================
# LatCoder
easyocr
opencv-python
open_clip_torch

# UICopilot uses Pix2Struct from transformers (already included above).

# Design2Code fine-grained metrics
# After installing Python packages, install Playwright's browser runtime with:
#   python -m playwright install chromium
# On Linux servers, also install Chromium system dependencies with:
#   python -m playwright install-deps chromium
openai-clip==1.0.1
colormath

# ============================================================
# Optional dev/notebook tooling
# ============================================================
ipykernel
ipywidgets
